# swagger/default.yaml
openapi: '3.0.0'
info:
  title: rakenduste-programmeerimine
  description: ""
  version: "1.0.0"
servers:
  - url: localhost:3000/api/v1
security:
  - bearerAuth: []
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      summary: Returns an user object
      security:
        - bearerAuth: []
      parameters:
         - $ref: "#/components/parameters/userIdParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        403:
          description: Forbidden
        500:
          description: Error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        createdAt:
          type: string
        cart:
          type: array
        items:
            type: string
        email:
          type: string
        username:
          type: string
  parameters:
    userIdParam:
      in: path
      name: userId
      schema:
        type: string
      required:  true
...